fixe kcp test warning

+++ kcp_test/client.cpp
@@ -100,7 +100,7 @@ void Client::input()
 void Client::run()
 {
     char buff[BUFFER_SIZE];
-    uint32_t next = 0;
+    int next = 0;
     uint32_t sumrtt = 0;
     uint32_t count = 0;
     uint32_t maxrtt = 0;
@@ -125,7 +125,7 @@ void Client::run()
         IUINT32 ts = *(IUINT32 *)( buff + 4 );
         IUINT32 rtt = util::iclock() - ts;
 
-        if ( sn != next ) {
+        if ( sn != (uint32_t)next ) {
             printf( "ERROR sn %d<-> next=%d\n", sn, next );
             is_running = false;
             break;
